Add styling to search form in homepage
Styling was added to the search from in the homepage. There is a slight issue where if you resize the window too small the search form will block the header buttons. I also removed a ton of divs in the code surrounding various items in the header because they did nothing.